first stage p2p v1

+++ database/tables/tickets.table.sql
@@ -5,8 +5,8 @@ CREATE TYPE ticket_status AS ENUM ('pending', 'precompleted', 'completed', 'pref
 CREATE TABLE tickets (
     ticket_id BIGSERIAL UNIQUE NOT NULL PRIMARY KEY,
     owner BIGINT REFERENCES usuarios(id) NOT NULL,
-    amount NUMERIC NOT NULL,
-    remain NUMERIC NOT NULL CHECK(remain <= amount),
+    amount DOUBLE PRECISION NOT NULL,
+    remain DOUBLE PRECISION NOT NULL CHECK(remain <= amount),
     created_at timestamptz NOT NULL,
     status TICKET_STATUS DEFAULT('pending') NOT NULL,
     type TICKET_TYPE NOT NULL
